chore(suite): get rid of useDaysTo hook
